fixed bug with live update disabling
