+++ OsmAnd/src/net/osmand/plus/track/fragments/SplitIntervalBottomSheet.java
@@ -128,9 +128,7 @@ public class SplitIntervalBottomSheet extends MenuBottomSheetDialogFragment {
 		TextRadioItem distance = createRadioButton(GpxSplitType.DISTANCE, R.string.distance);
 
 		GPXTrackAnalysis analysis = selectedGpxFile.getTrackAnalysisToDisplay(app);
-		if (analysis.timeSpan == 0){
-			time.setEnabled(false);
-		}
+		time.setEnabled(analysis.timeSpan > 0);
 
 		TextToggleButton radioGroup = new TextToggleButton(app, buttonsContainer, nightMode);
 		radioGroup.setItems(none, time, distance);
