+++ OsmAnd/src/net/osmand/plus/widgets/multistatetoggle/TextToggleButton.java
@@ -41,7 +41,6 @@ public class TextToggleButton extends MultiStateToggleButton<TextRadioItem> {
 	public static class TextRadioItem extends RadioItem {
 
 		private final String title;
-		private boolean isEnabled = true;
 
 		public TextRadioItem(String title) {
 			this.title = title;
@@ -50,14 +49,5 @@ public class TextToggleButton extends MultiStateToggleButton<TextRadioItem> {
 		public String getTitle() {
 			return title;
 		}
-
-		public boolean isEnabled() {
-			return isEnabled;
-		}
-
-		public void setEnabled(boolean enabled) {
-			isEnabled = enabled;
-		}
 	}
-
 }
